requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy==1.21.5
pandas==1.3.5
scikit-learn==1.0.2
#xgboost
scipy==1.4.1
statsmodels==0.11.0

# tests/linter
#black
#coverage
#flake8
#pytest
#yapf

# API
gcsfs
google-cloud-storage==1.18.1
mlflow
s3fs
pmdarima

# utilities
six>=1.14
joblib
memoized-property
termcolor

# API
fastapi
uvicorn
